Modularize `iced_wgpu`
