+++ wgpu/src/renderer/radio.rs
@@ -1,0 +1,17 @@
+use crate::{Primitive, Renderer};
+use iced_native::{radio, Layout, Node, Point, Radio, Style};
+
+impl radio::Renderer for Renderer {
+    fn node<Message>(&mut self, _checkbox: &Radio<Message>) -> Node {
+        Node::new(Style::default())
+    }
+
+    fn draw<Message>(
+        &mut self,
+        _radio: &Radio<Message>,
+        _layout: Layout<'_>,
+        _cursor_position: Point,
+    ) -> Self::Primitive {
+        Primitive::None
+    }
+}
